Add ability to mappe individual LUNs to ACLs
WARNING: this is breaking change as the format of 'initiators' has
changed.

Motivation for this change was https://github.com/OndrejHome/ansible.targetcli-modules/issues/5

To be able to use this new format the relevant
OndrejHome.targetcli-modules are needed.

Known issues: the cluster configuration use case is broken at this time.